test(engine): positively discriminate the assimilate target filter
Tests 3a and 3b asserted only that an illegal candidate stayed in its
graveyard, which is equally consistent with "the filter rejected it" and
"the ETB production never fired at all". Seed a legal AND an illegal
candidate in one fixture and declare the illegal one FIRST, so the filter
is observable across all three reachable states:

  correct     the legal set is a singleton, so
              auto_select_targets_for_ability assigns it and no prompt is
              raised; the legal card enters, the illegal one stays
  over-match  two legal candidates, auto-selection declines, a required
              slot IS created, and pick_slot_target fills it from the
              DECLARED order -- taking the bait, flipping both legs
  under-match empty legal set, so CR 603.3d removes the trigger before
              any slot exists and the POSITIVE leg fails

Declaration order is a regression-only instrument: it converts the
over-match case from pick_slot_target's no-declared-target panic into a
clean assertion failure. Each test holds two filter legs constant and
varies one, so 3a isolates the Creature type leg and 3b the ownership leg
(CR 109.4: a graveyard card has no controller, so "an opponent's
graveyard" rides as ownership). InZone { Graveyard } is isolated by
neither and keeps only its parser-level coverage.

Both tests are now revert-failing. 3b previously passed with the lowering
reverted, because nothing moved at all, and its doc comment claimed only
forward-guard status; that claim is updated rather than left stale.

Correct Recruit's mis-cited keyword-action rule. CR 701.9 is Discard;
Recruit is CR 701.70a. The "standalone keyword action" comment is purely
a 701.70a claim, so it is replaced. The discard-result comment genuinely
involves both rules -- the discard is 701.9a, the nonland contingency is
701.70a -- so it gains 701.70a instead. Two further Recruit-tagged 701.9a
sites are deliberately left alone: they assert snapshot and provenance
semantics already covered by 701.9a + 608.2c + 400.7, and assert nothing
that 701.70a supplies.

Relocate the parser-only fail-closed test beside its sibling shape test
in the parser module, and drop the two imports it was the last user of.

Replace the ephemeral BASE_SHA run identifier in three doc comments with
a stable PR reference. BASE_SHA is an implementation-run artifact whose
meaning changes every run and is undefined to a reader of the repository.
